Scoring and Game Over
Final commit, adds scoring and game over screen for the player. There
are glitches such as bullets being produced from non-existent enemies
and enemies shooting each other out, but the games basic mechanics work
and it's more of a learning project. I do not plan to further update
this project, but anyone is welcome to fork it, make it there own, and
fix it up.